+++ definitions/consts/shared-config.const.ts
@@ -1,0 +1,5 @@
+export const SHARED_CONFIG: {
+  cloudRegion: 'us-central1' | 'us-east1' | 'us-east4' | 'europe-west1' | 'europe-west2' | 'asia-east2' | 'asia-northeast1';
+} = {
+	cloudRegion: 'us-central1',
+};