+++ functions/src/enums/firestore-collections.enum.ts
@@ -1,5 +1,0 @@
-export enum FirestoreCollection {
-  Modules = 'modules',
-  Settings = 'settings',
-  Users = 'users'
-}