+++ functions/src/triggers/file-deleted.ts
@@ -1,13 +1,13 @@
 import {Storage} from '@google-cloud/storage';
 import * as functions from 'firebase-functions';
 import {basename, dirname, join} from 'path';
-import {STATIC_CONFIG} from '../consts/static-config.const';
 import {unpackGenerateImageString} from '../utils/unpack-generate-image-string';
 import {firestore} from 'firebase-admin';
 import {ObjectMetadata} from 'firebase-functions/lib/providers/storage';
+import {SHARED_CONFIG} from 'definitions';
 
 export const fileDeleted = functions
-  .region(STATIC_CONFIG.cloudRegion)
+  .region(SHARED_CONFIG.cloudRegion)
   .storage
   .object()
   .onDelete(async (data: ObjectMetadata) => {