+++ functions/src/triggers/user-document-updated.ts
@@ -1,5 +1,4 @@
-import {STATIC_CONFIG} from '../consts/static-config.const';
-import {FirestoreCollection} from '../enums/firestore-collections.enum';
+import {SHARED_CONFIG, Collections} from 'definitions';
 import * as functions from 'firebase-functions';
 import {auth} from 'firebase-admin';
 
@@ -8,9 +7,9 @@ import {auth} from 'firebase-admin';
  * the users role changes in firestore
  */
 export const userDocumentUpdated = functions
-  .region(STATIC_CONFIG.cloudRegion)
+  .region(SHARED_CONFIG.cloudRegion)
   .firestore
-  .document(`${FirestoreCollection.Users}/{documentId}`)
+  .document(`${Collections.Users}/{documentId}`)
   .onUpdate(async change => {
     const after: any = change.after.data();
     const before: any = change.before.data();